requirements.txt
pyyaml~=6.0
selenium~=4.1.0
loguru~=0.6.0
requests~=2.27.1
webdriver_manager>=3.5.2